# docs/public/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: API Support
  description: 
    GO API template made using common tecnologies.
    With user management demonstration built in
  license:
    name: GNU GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.html
  termsOfService: "not available yet"
  title: GO API template
  version: "1.0"
servers:
  - url: https://api.kaomoji-db.com:{port}/{basePath}
    description: The production API server
    variables:
      port:
        enum:
          - "80"
          - "443"
        default: "443"
      basePath:
        enum:
          - v1
          - prev
        # open meaning there is the opportunity to use special base paths as assigned by the provider, default is `v1`
        default: v1
  - url: http://localhost:{port}/{basePath}
    description: The local development API
    variables:
      port:
        enum:
          - '3000'
          - '443'
        default: '3000'
      basePath:
        enum:
          - 'v1'
          - 'prev'
        default: 'v1'
components:
  securitySchemes:
    bearer:
        type: "http"
        scheme: "bearer"
        bearerFormat: "JWT"
...
  responses:
    "500":
      description: Internal Server Error
      schema:
        $ref: "#/components/schemas/error"
    "400":
      description: Bad Request
      schema:
        $ref: "#/components/schemas/error"
    "401":
      description: Unauthorized
      schema:
        $ref: "#/components/schemas/error"
...
  schemas:
    token.response:
      type: object
      properties:
        status:
          type: string
          description: status string.
        message:
          type: string
          description: Specific description of the action
        token:
          type: string
          description: Raw bearer token in JWT format
...
    error:
      type: object
      properties:
        status:
          type: string
          description: status string, 'sucess' or 'error'
        message:
          type: string
          description: Specific description of the error
        data:
          type: string
          description: Data related to the error
paths:
...
  /auth/drop:
    get:
      summary: drops (logout) bearer token
      description: Blocks a given token, if none is provided, the one used to acess this route willbe blocked.
      requestBody:
        description: token json field
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: user's token to be blocked
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/components/schemas/token.response"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
      security:
        - bearer: []
...